requirements.txt
flask
psycopg2-binary
secrets
dotenv
openai==1.69.1